+++ app.vue
@@ -1,5 +1,5 @@
 <template>
-  <div>
+  <div class="layout">
     <NuxtPage />
   </div>
   <FooterApiFooter />
@@ -8,7 +8,7 @@
 <script lang="ts" setup>
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { faUser } from '@fortawesome/free-regular-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
+// import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 
 library.add(faUser)
 
@@ -19,8 +19,14 @@ useHead(() => ({
     { rel: 'stylesheet', href: 'https://site-assets.fontawesome.com/releases/v6.4.0/css/all.css' },
     { rel: 'stylesheet', href: 'https://site-assets.fontawesome.com/releases/v6.4.0/css/sharp-solid.css' },
     { rel: 'stylesheet', href: 'https://site-assets.fontawesome.com/releases/v6.4.0/css/sharp-regular.css' },
-    { rel: 'stylesheet', href: 'https://site-assets.fontawesome.com/releases/v6.4.0/css/sharp-light.css' }
+    { rel: 'stylesheet', href: 'https://site-assets.fontawesome.com/releases/v6.4.0/css/sharp-light.css' },
   ]
 }))
 
 </script>
+
+<style lang='scss' scoped>
+.layout {
+  height: 85vh;
+}
+</style>