+++ components/footer/api-footer.vue
@@ -1,5 +1,7 @@
 <template>
-  <FooterApiFooterCopywrite />
+  <div class="footer">
+    <FooterApiFooterCopywrite />
+  </div>
 </template>
 
 <script setup lang='ts'>
@@ -7,5 +9,7 @@
 </script>
 
 <style lang='scss' scoped>
-  
+  .footer {
+    height: 15%;
+  }
 </style>